Jest testing framework
sec. 16, video 137

$ npm i jest@23.6.0 --save-dev

The .test extension makes sure that jest knows
the file contains test cases that we want to run.

Jest provides the 'test' function as a global in our
test suite files so we don't have to require it manually.

If the function throws an error then the test case is considered
a failure.